change setId() parameter to type long

+++ src/main/java/numberGuess/Player.java
@@ -19,7 +19,7 @@ public class Player {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
